+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 <!-- _No changes yet._ -->
 
 - feat: update Sound, Mic & HackerNews widgets positioning to support dynamic bar position adjustments
+- refactor: improve widget reordering widgets in the bar editor
 
 ## v1.2.0 - 2026-02-09